+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### v2.4.2
 
 - Add Github Discussions entrance
+- Fix Vector drawable on Android 5
 - Fix Settings ModalDrawerSheet inserts [#212](https://github.com/hushenghao/AndroidEasterEggs/issues/212)
 
 ### v2.4.1 (2024-03-14)